+++ src/main/java/com/chat_server/chatmessage/repository/ChatMessageRepository.java
@@ -1,0 +1,8 @@
+package com.chat_server.chatmessage.repository;
+
+import com.chat_server.chatmessage.entity.ChatMessage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+
+}
